refactor for main
