+++ cmd/packageloader/loader/cover.out
@@ -3,21 +3,21 @@ github.com/skycoin/cx/cmd/packageloader/packageloader.go:38.51,39.28 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:44.2,44.14 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:39.28,40.22 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:40.22,42.4 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:47.13,48.39 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:53.2,56.86 3 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:65.2,65.16 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:69.2,69.37 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:72.2,73.33 2 0
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:47.13,48.39 1 1
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:53.2,56.86 3 1
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:65.2,65.16 1 1
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:69.2,69.37 1 1
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:72.2,73.33 2 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:48.39,50.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:50.8,52.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:56.86,57.17 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:60.3,60.19 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:63.3,63.13 1 0
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:50.8,52.3 1 1
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:56.86,57.17 1 1
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:60.3,60.19 1 1
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:63.3,63.13 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:57.17,59.4 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:60.19,62.4 1 0
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:60.19,62.4 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:65.16,67.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:69.37,71.3 1 0
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:73.33,75.3 1 0
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:69.37,71.3 1 1
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:73.33,75.3 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:78.60,79.31 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:82.2,83.50 2 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:86.2,90.16 5 1
@@ -26,7 +26,7 @@ github.com/skycoin/cx/cmd/packageloader/packageloader.go:99.2,99.24 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:110.2,110.23 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:123.2,126.31 4 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:79.31,81.3 1 1
-github.com/skycoin/cx/cmd/packageloader/packageloader.go:83.50,85.3 1 0
+github.com/skycoin/cx/cmd/packageloader/packageloader.go:83.50,85.3 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:90.16,92.3 1 0
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:93.33,94.55 1 1
 github.com/skycoin/cx/cmd/packageloader/packageloader.go:97.3,97.40 1 1
